# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Trosc API Documentation",
    "version": "1.0.0",
    "description": "Official API documentation for the Trosc platform.",
    "contact": {
      "name": "Basem Esam",
      "url": "https://github.com/basem3sam"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/v1",
      "description": "development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UserCreate": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password",
          "passwordConfirm"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50,
            "example": "Basem Esam"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "basem@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8,
            "example": "StrongP@ssw0rd123"
          },
          "passwordConfirm": {
            "type": "string",
            "format": "password",
            "example": "StrongP@ssw0rd123"
          },
          "photo": {
            "type": "string",
            "example": "profile.jpg"
          },
          "bio": {
            "type": "string",
            "maxLength": 500,
            "example": "Backend Engineer | ICPC Competitor"
          }
        }
      },
      "AuthLogin": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "basem@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "StrongP@ssw0rd123"
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/UserBase"
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Access token is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardError"
            }
          }
        }
      },
      "ValidationError": {
        "description": "Validation failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/StandardError"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/users/signup": {
      "post": {
        "operationId": "signup",
        "summary": "Register a new user account",
        "description": "Creates a new user account with student role by default",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "409": {
            "description": "Email already registered"
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "operationId": "login",
        "summary": "Authenticate user and return JWT token",
        "description": "Verifies user credentials and returns JWT token",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "operationId": "logout",
        "summary": "Log out current user",
        "description": "Clears authentication token",
        "tags": [
          "Auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Logged out successfully"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/forgotPassword": {
      "post": {
        "operationId": "forgotPassword",
        "summary": "Send password reset token to user's email",
        "description": "Generates and sends password reset token via email",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "basem@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset email sent"
          },
          "404": {
            "description": "No user found with provided email"
          }
        }
      }
    },
    "/users/resetPassword/{token}": {
      "patch": {
        "operationId": "resetPassword",
        "summary": "Reset password using token from email",
        "description": "Resets user password using valid reset token",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordReset"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully"
          },
          "400": {
            "description": "Invalid or expired token"
          }
        }
      }
    },
    "/users/updateMyPassword": {
      "patch": {
        "operationId": "updateMyPassword",
        "summary": "Update current user's password",
        "description": "Change password for authenticated user",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated successfully"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/me/enrollments": {
      "get": {
        "operationId": "getMyEnrollments",
        "summary": "Get my track and course enrollments",
        "description": "Returns all tracks and courses the current user is enrolled in",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Enrollments retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "tracks": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Track"
                          }
                        },
                        "courses": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Course"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "operationId": "getMyProfile",
        "summary": "Get current user's profile",
        "description": "Returns profile of authenticated user",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/UserBase"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/updateMe": {
      "patch": {
        "operationId": "updateMyProfile",
        "summary": "Update current user's profile information",
        "description": "Update user profile (name, email, photo, bio)",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Basem Updated"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "basem.updated@example.com"
                  },
                  "photo": {
                    "type": "string",
                    "example": "new-avatar.jpg"
                  },
                  "bio": {
                    "type": "string",
                    "example": "Senior Backend Engineer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "$ref": "#/components/schemas/UserBase"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/deleteMe": {
      "delete": {
        "operationId": "deleteMyAccount",
        "summary": "Deactivate current user account (soft delete)",
        "description": "Soft deletes user account by setting active=false",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Account deactivated"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users": {
      "get": {
        "operationId": "getAllUsers",
        "summary": "Get all users (admin only)",
        "description": "Retrieves list of all active users",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of users retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "operationId": "createUser",
        "summary": "Create a new user (admin only)",
        "description": "Admin endpoint to create users with specific roles",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserAdminCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/users/bulk": {
      "post": {
        "operationId": "bulkUserAction",
        "summary": "Bulk user actions (admin only)",
        "description": "Activate, deactivate, or delete multiple users at once",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userIds",
                  "action"
                ],
                "properties": {
                  "userIds": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "507f1f77bcf86cd799439011"
                    },
                    "example": [
                      "507f1f77bcf86cd799439011",
                      "507f1f77bcf86cd799439012"
                    ]
                  },
                  "action": {
                    "type": "string",
                    "enum": [
                      "activate",
                      "deactivate",
                      "delete"
                    ],
                    "example": "activate"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk action completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Bulk activate completed successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    }
  },
...
}